Move get_mean_elements function to separate module to avoid circular import
